Reorganize code for loading certificates
